Telemetry saving design proposals
Implementation of a proof of concept function that saves telemetry increasing memory adress and keeps track of where to save the data after MCU resets. Two designs are proposed one where this value is saved at the beggining of the reserved space and one where it is stored with the rest of the data.Also inclusion of a page removal function. NEEDS REAL TESTING, COMPILES